chore: rename otel tag and backend
